Allow vulnerabilities to be whitelisted by feature name
This change allows specific CVEs to be whitelisted when they affect a
specific feature. This is useful when a feature embeds a vulnerable
component, but uses it in a way that is not affected by the
vulnerability - and the user wants to ensure the CVE does not affect
other features (such as the OS Distribution's own package that
distributes the component).

See https://security-tracker.debian.org/tracker/CVE-2019-8457 for a
motivating example. Users may want to whitelist CVE-2019-8457 in the
context of the db5.3 package, but still be confident that sqlite3 has
been patched.